Restructuring objects
