+++ youtube-dl-gui/Classes/Formats.cs
@@ -48,9 +48,68 @@ internal class Formats {
     public const string AllKnownAudioFormats =
         "All known audio formats|*.aac;*.aiff;*.aif;*.aifc;*.flac;*.m4a;*.mp4;*.mp2;*.mp3;*.oga;*.ogg;*.opus;*.wav";
 
+
+
+    #region Video Arrays
+    public static readonly string[] VideoQualityArray = { "best",
+                                                 "4320p60", "4320p", // 1
+                                                 "2160p60", "2160p", // 3
+                                                 "1440p60", "1440p", // 5
+                                                 "1080p60", "1080p", // 7
+                                                 "720p60", "720p",   // 9
+                                                 "480p",
+                                                 "360p",
+                                                 "240p",
+                                                 "144p",
+                                                 "worst" };
+    public static readonly string[] VideoFormatsNamesArray = { "best",
+                                                      "avi",
+                                                      "flv",
+                                                      "mkv",
+                                                      "mp4",
+                                                      "ogg",
+                                                      "webm"
+                                                    };
+    public static readonly string[] VideoFormatsArray = { " --recode-video avi",
+                                                 " --recode-video flv",
+                                                 " --merge-output-format mkv",//" --recode-video mkv", //" --merge-output-format mkv",
+                                                 "",//" --recode-video mp4", //" --merge-output-format mp4",
+                                                 " --recode-video ogg",
+                                                 " --recode-video webm" //" --merge-output-format webm"
+                                               };
+
+    //mp4|flv|ogg|webm|mkv|avi
+
+    #endregion
+
+    #region Audio Arrays
+    public static readonly string[] AudioQualityNamesArray = { "best",
+                                                      "320k",
+                                                      "256k",
+                                                      "224k",
+                                                      "192k",
+                                                      "160k",
+                                                      "128k",
+                                                      "96k",
+                                                      "64k",
+                                                      "32k",
+                                                      "16k",
+                                                      "worst"
+                                              };
+    public static readonly string[] AudioFormatsArray = { "best",
+                                                 "aac",
+                                                 "flac",
+                                                 "mp3",
+                                                 "m4a",
+                                                 "opus",
+                                                 "vorbis",
+                                                 "wav"
+                                             };
+
     public static readonly string[] VbrQualities = new string[] {
         "0", "1", "2", "3", "4", "5", "6", "7", "8", "9", "10"
     };
+    #endregion
 
     // Could this be replaced with video formats? Because they could contain video, but don't necessarily need to.
     /// <summary>
@@ -100,5 +159,85 @@ internal class Formats {
     public static string JoinFormats(string[] FormatsArray) {
         var Formats = FormatsArray.Where(ArrayIndex => !string.IsNullOrWhiteSpace(ArrayIndex));
         return Formats.Count() > 0 ? string.Join("|", Formats) : AllFiles;
+    }
+
+    public static string GetVideoQualityArgs(VideoQualityType Quality) {
+        return Quality switch {
+            VideoQualityType.q4320p60 => " -f \"bestvideo[height<=4320][fps>=48]+bestaudio/best\"",
+            VideoQualityType.q4320p => " -f \"bestvideo[height<=4320][fps<=32]+bestaudio/best\"",
+            VideoQualityType.q2160p60 => " -f \"bestvideo[height<=2160][fps>=48]+bestaudio/best\"",
+            VideoQualityType.q2160p => " -f \"bestvideo[height<=2160][fps<=32]+bestaudio/best\"",
+            VideoQualityType.q1440p60 => " -f \"bestvideo[height<=1440][fps>=48]+bestaudio/best\"",
+            VideoQualityType.q1440p => " -f \"bestvideo[height<=1440][fps<=32]+bestaudio/best\"",
+            VideoQualityType.q1080p60 => " -f \"bestvideo[height<=1080][fps>=48]+bestaudio/best\"",
+            VideoQualityType.q1080p => " -f \"bestvideo[height<=1080][fps<=32]+bestaudio/best\"",
+            VideoQualityType.q720p60 => " -f \"bestvideo[height<=720][fps>=48]+bestaudio/best\"",
+            VideoQualityType.q720p => " -f \"bestvideo[height<=720][fps<=32]+bestaudio/best\"",
+            VideoQualityType.q480p => " -f \"bestvideo[height<=480]+bestaudio/best\"",
+            VideoQualityType.q360p => " -f \"bestvideo[height<=360]+bestaudio/best\"",
+            VideoQualityType.q240p => " -f \"bestvideo[height<=240]+bestaudio/best\"",
+            VideoQualityType.q144p => " -f \"bestvideo[height<=144]+bestaudio/best\"",
+            VideoQualityType.worst => " -f \"worstvideo+worstaudio/worst\"",
+            _ => "",
+        };
+    }
+    public static string GetVideoQualityArgsNoSound(VideoQualityType Quality) {
+        return Quality switch {
+            VideoQualityType.q4320p60 => " -f \"bestvideo[height<=4320][fps>=48]/best\"",
+            VideoQualityType.q4320p => " -f \"bestvideo[height<=4320][fps<=32]/best\"",
+            VideoQualityType.q2160p60 => " -f \"bestvideo[height<=2160][fps>=48]/best\"",
+            VideoQualityType.q2160p => " -f \"bestvideo[height<=2160][fps<=32]/best\"",
+            VideoQualityType.q1440p60 => " -f \"bestvideo[height<=1440][fps>=48]/best\"",
+            VideoQualityType.q1440p => " -f \"bestvideo[height<=1440][fps<=32]/best\"",
+            VideoQualityType.q1080p60 => " -f \"bestvideo[height<=1080][fps>=48]/best\"",
+            VideoQualityType.q1080p => " -f \"bestvideo[height<=1080][fps<=32]/best\"",
+            VideoQualityType.q720p60 => " -f \"bestvideo[height<=720][fps>=48]/best\"",
+            VideoQualityType.q720p => " -f \"bestvideo[height<=720][fps<=32]/best\"",
+            VideoQualityType.q480p => " -f \"bestvideo[height<=480]/best\"",
+            VideoQualityType.q360p => " -f \"bestvideo[height<=360]/best\"",
+            VideoQualityType.q240p => " -f \"bestvideo[height<=240]/best\"",
+            VideoQualityType.q144p => " -f \"bestvideo[height<=144]/best\"",
+            VideoQualityType.worst => " -f \"worstvideo\"",
+            _ => "",
+        };
+    }
+    public static string GetVideoRecodeInfo(VideoFormatType Format) {
+        return Format switch {
+            VideoFormatType.avi => " --recode-video avi",
+            VideoFormatType.flv => " --recode-video flv",
+            VideoFormatType.mkv => " --merge-output-format mkv",
+            VideoFormatType.ogg => " --recode-video ogg",
+            VideoFormatType.webm => " --merge-output-format webm",
+            _ => string.Empty,
+        };
+    }
+    public static string GetAudioFormat(AudioFormatType Format) {
+        return Format switch {
+            AudioFormatType.aac => "aac",
+            AudioFormatType.flac => "flac",
+            AudioFormatType.m4a => "m4a",
+            AudioFormatType.mp3 => "mp3",
+            AudioFormatType.opus => "opus",
+            AudioFormatType.vorbis => "vorbis",
+            AudioFormatType.wav => "wav",
+            _ => "best",
+        };
+    }
+    public static string GetAudioQuality(AudioCBRQualityType Quality) {
+        return Quality switch {
+            AudioCBRQualityType.q320k => "320k",
+            AudioCBRQualityType.q256k => "256k",
+            AudioCBRQualityType.q244k => "224k",
+            AudioCBRQualityType.q192k => "192k",
+            AudioCBRQualityType.q160k => "160k",
+            AudioCBRQualityType.q128k => "128k",
+            AudioCBRQualityType.q96k => "96k",
+            AudioCBRQualityType.q64k => "64k",
+            AudioCBRQualityType.q32k => "32k",
+            AudioCBRQualityType.q16k => "16k",
+            AudioCBRQualityType.q8k => "8k",
+            AudioCBRQualityType.q4k => "4k",
+            _ => "best",
+        };
     }
 }
