+++ youtube-dl-gui/Controls/Enums/ButtonAlignment.cs
@@ -1,0 +1,15 @@
+﻿namespace murrty.controls;
+
+/// <summary>
+/// An enumeration of the alignments that the button can be aligned to.
+/// </summary>
+public enum ButtonAlignment {
+    /// <summary>
+    /// The Button will appear on the left side of the TextBox.
+    /// </summary>
+    Left,
+    /// <summary>
+    /// The Button will appear on the right side of the TextBox. Default value.
+    /// </summary>
+    Right,
+}
