+++ youtube-dl-gui/Resources/BuildDate.txt
@@ -1,1 +1,1 @@
-0000-00-00
+2022-09-05
