+++ youtube-dl-gui/Resources/StatusIcons/download.png
