+++ youtube-dl-gui/Resources/StatusIcons/error.png
