+++ youtube-dl-gui/Resources/StatusIcons/finished.png
