+++ youtube-dl-gui/Resources/StatusIcons/waiting.png
